requirements.txt
python-dotenv==0.21.0
neo4j~=5.15.0
pandas==2.1.4
Flask~=2.2.2
Werkzeug~=3.0.1
